+++ src/WebRadio_Radiko.hpp
@@ -339,7 +339,7 @@ class Radiko : public WebRadio {
       areaFree = false;
       
       if(!bufferSize)
-        bufferSize = std::max(6UL * 1024UL, std::min(256UL * 1024UL, (uint32_t)heap_caps_get_free_size(MALLOC_CAP_SPIRAM)));
+        bufferSize = std::max((uint32_t)(6 * 1024), std::min((uint32_t)(256 * 1024), (uint32_t)heap_caps_get_free_size(MALLOC_CAP_SPIRAM)));
       
       buffer = new AudioFileSourceHLS(bufferSize);
